refactor main game loop to allow clients to control game logic
