+++ src/test/scala/game/BambooEngineSpec.scala
@@ -5,19 +5,19 @@ import org.scalatest.matchers._
 
 import scala.collection.mutable.Stack
 
-class RunnerSpec extends AnyFlatSpec with should.Matchers {
+class BambooEngineSpec extends AnyFlatSpec with should.Matchers {
 
   "A Stack" should "pop values in last-in-first-out order" in {
     val stack = new Stack[Int]
     stack.push(1)
     stack.push(2)
-    stack.pop() should be (2)
-    stack.pop() should be (1)
+    stack.pop() should be(2)
+    stack.pop() should be(1)
   }
 
   it should "throw NoSuchElementException if an empty stack is popped" in {
     val emptyStack = new Stack[Int]
-    a [NoSuchElementException] should be thrownBy {
+    a[NoSuchElementException] should be thrownBy {
       emptyStack.pop()
     }
   }
